Update auth tests to work
